Update PumpkinSquire.cs
lots of buffs and changes. damage and speed are both buffed from 14 > 15, and 40 > 32 respectively. whip is 25% longer, big pumpkin damage has also been buffed to deal x2 damage rather than x1.5.

if i was more familiar with coding for this mod, i would have also made the pumpkin bomb explode after a while and release smaller ones but as i said im not very familiar with the codebase for this mod haha